+++ core/src/main/java/io/kestra/core/services/SkipExecutionService.java
@@ -1,4 +1,4 @@
-package io.kestra.core.runners;
+package io.kestra.core.services;
 
 import com.google.common.annotations.VisibleForTesting;
 import io.kestra.core.models.executions.Execution;
